+++ src/index.js
@@ -1,7 +1,7 @@
 import "./index.css";
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./pages/App/App";
+import App from "./pages/App";
 import { BrowserRouter } from "react-router-dom";
 
 ReactDOM.render(
