+++ src/pages/App/index.js
@@ -1,10 +1,10 @@
-import "./App.css";
+import "./index.css";
 import { useState } from "react";
 import { URL } from "../../constants";
 import { Header } from "../../components";
-import CardAddition from "../CardAddition/CardAddition";
-import CompleteCardAddition from "../CompleteCardAddition/CompleteCardAddition";
-import MyCardList from "../MyCardList/MyCardList";
+import CardAddition from "../CardAddition";
+import CompleteCardAddition from "../CompleteCardAddition";
+import MyCardList from "../MyCardList";
 import { Route, Switch, useHistory, useLocation } from "react-router";
 import { QUERY_STRING_KEY } from "../../constants.js";
