+++ src/pages/MyCardList/index.css
@@ -1,0 +1,3 @@
+ul {
+  list-style: none;
+}
